+++ inst/include/mlpack/core/metrics/mahalanobis_distance.hpp
@@ -4,7 +4,7 @@
  *
  * The Mahalanobis distance.
  *
- * This file is part of MLPACK 1.0.9.
+ * This file is part of MLPACK 1.0.10.
  *
  * MLPACK is free software: you can redistribute it and/or modify it under the
  * terms of the GNU Lesser General Public License as published by the Free
@@ -73,7 +73,7 @@ class MahalanobisDistance
    *
    * @param dimensionality Dimesnsionality of the covariance matrix.
    */
-  MahalanobisDistance(const long dimensionality) :
+  MahalanobisDistance(const size_t dimensionality) :
       covariance(arma::eye<arma::mat>(dimensionality, dimensionality)) { }
 
   /**
