+++ inst/include/mlpack/core/optimizers/sgd/sgd_impl.hpp
@@ -4,7 +4,7 @@
  *
  * Implementation of stochastic gradient descent.
  *
- * This file is part of MLPACK 1.0.9.
+ * This file is part of MLPACK 1.0.10.
  *
  * MLPACK is free software: you can redistribute it and/or modify it under the
  * terms of the GNU Lesser General Public License as published by the Free
@@ -32,7 +32,7 @@ namespace optimization {
 template<typename DecomposableFunctionType>
 SGD<DecomposableFunctionType>::SGD(DecomposableFunctionType& function,
                                    const double stepSize,
-                                   const long maxIterations,
+                                   const size_t maxIterations,
                                    const double tolerance,
                                    const bool shuffle) :
     function(function),
@@ -47,7 +47,7 @@ template<typename DecomposableFunctionType>
 double SGD<DecomposableFunctionType>::Optimize(arma::mat& iterate)
 {
   // Find the number of functions to use.
-  const long numFunctions = function.NumFunctions();
+  const size_t numFunctions = function.NumFunctions();
 
   // This is used only if shuffle is true.
   arma::vec visitationOrder;
@@ -56,17 +56,17 @@ double SGD<DecomposableFunctionType>::Optimize(arma::mat& iterate)
         numFunctions));
 
   // To keep track of where we are and how things are going.
-  long currentFunction = 0;
+  size_t currentFunction = 0;
   double overallObjective = 0;
   double lastObjective = DBL_MAX;
 
   // Calculate the first objective function.
-  for (long i = 0; i < numFunctions; ++i)
+  for (size_t i = 0; i < numFunctions; ++i)
     overallObjective += function.Evaluate(iterate, i);
 
   // Now iterate!
   arma::mat gradient(iterate.n_rows, iterate.n_cols);
-  for (long i = 1; i != maxIterations; ++i, ++currentFunction)
+  for (size_t i = 1; i != maxIterations; ++i, ++currentFunction)
   {
     // Is this iteration the start of a sequence?
     if ((currentFunction % numFunctions) == 0)
